dateprojects/requirements.txt
pymysql>=1.1.0,<2.0.0
python-dotenv>=1.0.0,<2.0.0
requests>=2.32.0,<3.0.0
